requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
Pillow==9.3.0
pycodestyle==2.9.1
python-http-client==3.3.7
pytz==2022.6
sendgrid==6.9.7
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tomli==2.0.1
